requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
paho-mqtt==1.6.1
pydantic==1.10.7